Feat#8 pid control (#21)
* Pulled from Dev v9

* Speed controller closed loop controller implementation ready to test

* Updated sim to work with new control functions

* fixed merge?

* reset to 0 the drive and steer pid in module.java and reset pid in constantsWIP

* PID and FF values tunable through smartdashboard/advantagescope

* 2/1/2025
Programming
Protobot PID
Goals:
Develop the PID
Progreess :
Today the Programming Subgroup continue developing implenting the PID controller, they also changed the Pidgeon reedings and offsets so the front of the robot is actually the front. Also also  decided to not use the speed controller close conroller for the turn motors because they had been working for a long time with it and they haven't reached any success or adventage over the PID that we used in the past, they also fixed the wheels turning 180 degrees after his position crosses the 0 or 180 degrees by costraining the value between -pi and pi. At the same time they also update the absolute encoder offsets to determine if that changed over time
Next steps :
Start tunning the chassis by tomorrow 2/2/2025

* Commented out all instances of speed controller pid control for the turn motor

---------

Co-authored-by: Helen Ling <Helen.Ling@sisepuedefoundation.org>
Co-authored-by: Andres Lara <Andres.Lara@sisepuedefoundation.org>
Co-authored-by: Josue Raygoza <Josue.Raygoza@sisepuedefoundation.org>