Feat#41 pathfind closest element (#45)  UWU
* starting reef autoalignment command

* Branch pathfinding nearly complete, just need Branch poses

* i had angles, trig, degrees

* IT WORKSgit add .git add .git add . (in sim at least) BUT IT WORKSSSSSSSSSSS FHJIOASHJDOASJDOIKSAJDO WOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOO.
BRANCH poses now defined by translating AprilTag pose (should be more accuracy as it is from the AprilTag field layout for either field type.
The way we define our static commands makes it so that if we want a button that have multiple outputs (like lining up to the closest REEF BRANCH), then all the code must be in the return. This can be done by returning Commands.run() and in the runnable portion use a lambda function with '{]' (like this: ()-> {}) and put all code inside of there so it gets updated with every button press.

* WORK IRL TOOOOOOOOOOO

* testing new camera mounts

* A little clean up, ready for PR

---------

Co-authored-by: Josue Raygoza <Josue.Raygoza@sisepuedefoundation.org>